+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.FinalProject"
         tools:targetApi="31">
         <activity
-            android:name=".ui.leftbar.presenter.LeftBarActivity"
+            android:name=".ui.preLogin.presenter.PreLoginActivity"
             android:exported="false" />
         <activity
             android:name=".ui.home.presenter.HomeActivity"
@@ -27,7 +27,7 @@
             android:name=".ui.register.presenter.RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.preLogin.presenter.PreLoginActivity"
+            android:name=".ui.leftbar.presenter.LeftBarActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
