+++ app/src/main/java/com/example/finalproject/ui/commentsFragment/CommentsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.finalproject
+package com.example.finalproject.ui.commentsFragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,13 +6,11 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
+import com.example.finalproject.R
 import com.example.finalproject.databinding.FragmentCommentsBinding
-
 
 class CommentsFragment : Fragment() {
     private lateinit var binding: FragmentCommentsBinding
-
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
@@ -26,7 +24,6 @@ class CommentsFragment : Fragment() {
     ): View {
         binding = FragmentCommentsBinding.inflate(inflater, container, false)
         return binding.root
-
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -41,8 +38,5 @@ class CommentsFragment : Fragment() {
         binding.tvImagesFragment.setOnClickListener {
             findNavController().navigate(R.id.action_commentsFragment_to_imagesFragment)
         }
-
     }
-
-
 }
