+++ app/src/main/java/com/example/finalproject/ui/descriptionFragment/presenter/DescriptionFragment.kt
@@ -1,4 +1,4 @@
-package com.example.finalproject
+package com.example.finalproject.ui.descriptionFragment.presenter
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,15 +6,14 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
+import com.example.finalproject.R
 import com.example.finalproject.databinding.FragmentDescriptionBinding
 
 class DescriptionFragment : Fragment() {
     private lateinit var binding: FragmentDescriptionBinding
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-
         }
     }
 
@@ -38,9 +37,5 @@ class DescriptionFragment : Fragment() {
         binding.tvCommentsFragment.setOnClickListener {
             findNavController().navigate(R.id.action_descriptionFragment_to_commentsFragment)
         }
-
     }
-
-
-
 }
